+++ lib/guru/wrapper.dart
@@ -1,0 +1,13 @@
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'loginscreen.dart';
+import 'homepage.dart';
+
+class Wrapper extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    FirebaseUser firebaseUser = Provider.of<FirebaseUser>(context);
+    return (firebaseUser == null) ? Login() : Homepage();
+  }
+}
