--- lib/homepage.dart
@@ -1,57 +1,0 @@
-import 'package:flutter/material.dart';
-import 'tab_home.dart' as beranda;
-import 'tab_akun.dart' as profile;
-
-class Homepage extends StatefulWidget {
-  final String username;
-
-  Homepage(profile.Akun akun, {this.username});
-
-  @override
-  _HomepageState createState() => _HomepageState();
-}
-
-class _HomepageState extends State<Homepage>
-    with SingleTickerProviderStateMixin {
-  TabController controller;
-
-  @override
-  void initState() {
-    controller = new TabController(length: 2, vsync: this);
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: new AppBar(
-        backgroundColor: Colors.green,
-        title: new Text("Parenting Control"),
-        bottom: new TabBar(
-          controller: controller,
-          tabs: <Widget>[
-            new Tab(
-              icon: new Icon(Icons.home),
-            ),
-            new Tab(
-              icon: new Icon(Icons.person),
-            ),
-          ],
-        ),
-      ),
-      body: new TabBarView(
-        controller: controller,
-        children: <Widget>[
-          new beranda.Tabhome(),
-          new profile.Akun(),
-        ],
-      ),
-    );
-  }
-}
