+++ lib/welcomescreen.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:login/guru/loginscreen.dart';
+import 'package:login/verifikasi_guru.dart';
+import 'package:login/verifikasi_ortu.dart';
 
 class Welcomescreen extends StatefulWidget {
   @override
@@ -7,8 +8,12 @@ class Welcomescreen extends StatefulWidget {
 }
 
 class _WelcomescreenState extends State<Welcomescreen> {
-  void lewat() {
-    Navigator.push(context, MaterialPageRoute(builder: (context) => Login()));
+  guru() {
+    Navigator.push(context, MaterialPageRoute(builder: (context) => VerifikasiGuru()));
+  }
+
+  ortu() {
+    Navigator.push(context, MaterialPageRoute(builder: (context) => VerifikasiOrtu()));
   }
 
   @override
@@ -16,7 +21,6 @@ class _WelcomescreenState extends State<Welcomescreen> {
     return Scaffold(
       resizeToAvoidBottomPadding: false,
       body: Container(
-        //width: MediaQuery.of(context).size.width,
         padding: EdgeInsets.all(20.0),
         color: Colors.white,
         child: Column(
@@ -74,7 +78,7 @@ class _WelcomescreenState extends State<Welcomescreen> {
               child: Container(
                 height: 35.0,
                 child: InkWell(
-                  onTap: () => lewat(),
+                  onTap: () => guru(),
                   splashColor: Colors.white,
                   child: Center(
                     child: Text(
@@ -94,7 +98,7 @@ class _WelcomescreenState extends State<Welcomescreen> {
               child: Container(
                 height: 35.0,
                 child: InkWell(
-                  onTap: () => lewat(),
+                  onTap: () => ortu(),
                   splashColor: Colors.white,
                   child: Center(
                     child: Text(
